:recycle: refactoring in

+++ chrono.c
@@ -1,18 +1,18 @@
 
 #include "philosophers.h"
 
-int chrono_init(void)
+unsigned long int chrono_init(void)
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
 
-    int start_time = tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    unsigned long int start_time = tv.tv_sec * 1000 + tv.tv_usec / 1000;
     return (start_time);
 }
 
-int stamp_time(int start_time)
+unsigned long int stamp_time(unsigned long int start_time)
 {
-    int actual_time;
+    unsigned long int actual_time;
     struct timeval tv;
     gettimeofday(&tv, NULL);
